#18 refactor: DummyRepository 및 DummyDocumentRepository 위치 및 패키지 변경
- DummyRepository는 core.domain.dummy에서 health.repository로 이동
- DummyDocumentRepository가 새로 생성되었으며, health.repository에 위치
