+++ src/main/kotlin/com/millo/ollim/health/repository/DummyDocumentRepository.kt
@@ -1,0 +1,6 @@
+package com.millo.ollim.health.repository
+
+import com.millo.ollim.health.domain.DummyDocument
+import org.springframework.data.mongodb.repository.MongoRepository
+
+interface DummyDocumentRepository: MongoRepository<DummyDocument, String>
