+++ src/main/kotlin/com/millo/ollim/health/repository/DummyRepository.kt
@@ -1,5 +1,6 @@
-package com.millo.ollim.core.domain.dummy
+package com.millo.ollim.health.repository
 
+import com.millo.ollim.health.domain.DummyEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.rest.core.annotation.RepositoryRestResource
